+++ analysis/import/SourceMonitorImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcemonitor/ParserDialog.kt
@@ -1,28 +1,41 @@
 package de.maibornwolff.codecharta.importer.sourcemonitor
 
-import com.github.kinquirer.KInquirer
-import com.github.kinquirer.components.promptConfirm
-import com.github.kinquirer.components.promptInput
+import com.varabyte.kotter.runtime.RunScope
+import com.varabyte.kotter.runtime.Session
+import de.maibornwolff.codecharta.serialization.FileExtension
+import de.maibornwolff.codecharta.tools.inquirer.InputType
+import de.maibornwolff.codecharta.tools.inquirer.myPromptConfirm
+import de.maibornwolff.codecharta.tools.inquirer.myPromptDefaultFileFolderInput
+import de.maibornwolff.codecharta.tools.inquirer.myPromptInput
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
-import de.maibornwolff.codecharta.util.CSVFileAggregation.Companion.getInputFiles
 
 class ParserDialog {
     companion object : ParserDialogInterface {
-        override fun collectParserArgs(): List<String> {
-            val inputFileNames = getInputFiles("What is the SourceMonitor CSV file that has to be parsed?")
+        override fun collectParserArgs(session: Session): List<String> {
+            val inputFileName = session.myPromptDefaultFileFolderInput(
+                inputType = InputType.FILE,
+                fileExtensionList = listOf(FileExtension.CSV),
+                onInputReady = testCallback()
+            )
 
-            val outputFileName: String = KInquirer.promptInput(
+            val outputFileName: String = session.myPromptInput(
                 message = "What is the name of the output file?",
-                hint = "output.cc.json"
+                allowEmptyInput = true,
+                onInputReady = testCallback()
             )
 
-            val isCompressed = (outputFileName.isEmpty()) || KInquirer.promptConfirm(
-                message = "Do you want to compress the output file?", default = true
+            val isCompressed = (outputFileName.isEmpty()) || session.myPromptConfirm(
+                message = "Do you want to compress the output file?",
+                onInputReady = testCallback()
             )
 
-            return inputFileNames + listOfNotNull(
-                "--output-file=$outputFileName", if (isCompressed) null else "--not-compressed"
+            return listOfNotNull(
+                inputFileName,
+                "--output-file=$outputFileName",
+                if (isCompressed) null else "--not-compressed"
             )
         }
+
+        internal fun testCallback(): suspend RunScope.() -> Unit = {}
     }
 }