+++ analysis/tools/ccsh/src/main/kotlin/de/maibornwolff/codecharta/tools/ccsh/parser/ParserService.kt
@@ -1,14 +1,13 @@
 package de.maibornwolff.codecharta.tools.ccsh.parser
 
-import com.github.kinquirer.KInquirer
-import com.github.kinquirer.components.promptList
 import de.maibornwolff.codecharta.tools.ccsh.parser.repository.PicocliParserRepository
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
+import de.maibornwolff.codecharta.tools.interactiveparser.runInTerminalSession
 import picocli.CommandLine
 
 class ParserService {
     companion object {
-        private const val EXIT_CODE_PARSER_NOT_SUPPORTED = 42
+        internal const val EXIT_CODE_PARSER_NOT_SUPPORTED = 42
 
         fun getParserSuggestions(commandLine: CommandLine, parserRepository: PicocliParserRepository, inputFile: String): List<String> {
             val allParsers = parserRepository.getAllInteractiveParsers(commandLine)
@@ -27,25 +26,22 @@ class ParserService {
             for (selectedParser in selectedParsers) {
                 println(System.lineSeparator() + "Now configuring $selectedParser.")
                 val interactiveParser = parserRepository.getInteractiveParser(commandLine, selectedParser)
-                if (interactiveParser == null) {
-                    throw IllegalArgumentException("Tried to configure non existing parser!")
-                } else {
-                    val configuration = interactiveParser.getDialog().collectParserArgs()
-                    configuredParsers[selectedParser] = configuration
 
-                    println("You can run the same command again by using the following command line arguments:")
-                    println("ccsh " + selectedParser + " " + configuration.map { x -> '"' + x + '"' }.joinToString(" "))
-                }
+                requireNotNull(interactiveParser) { "Tried to configure non existing parser!" }
+
+                val configuration = runInTerminalSession { interactiveParser.getDialog().collectParserArgs(this) }
+
+                configuredParsers[selectedParser] = configuration
+
+                println("You can run the same command again by using the following command line arguments:")
+                println("ccsh " + selectedParser + " " + configuration.joinToString(" ") { x -> '"' + x + '"' })
             }
             return configuredParsers
         }
 
         fun selectParser(commandLine: CommandLine, parserRepository: PicocliParserRepository): String {
-            val selectedParser: String =
-                KInquirer.promptList(
-                    message = "Which parser do you want to execute?",
-                    choices = parserRepository.getInteractiveParserNamesWithDescription(commandLine)
-                )
+            val interactiveParserNames = parserRepository.getInteractiveParserNamesWithDescription(commandLine)
+            val selectedParser = runInTerminalSession { InteractiveDialog.askParserToExecute(this, interactiveParserNames) }
             return parserRepository.extractParserName(selectedParser)
         }
 
@@ -54,10 +50,11 @@ class ParserService {
             val parserObject = subCommand.commandSpec.userObject()
             val interactive = parserObject as? InteractiveParser
             return if (interactive != null) {
-                val collectedArgs = interactive.getDialog().collectParserArgs()
+                val collectedArgs = runInTerminalSession { interactive.getDialog().collectParserArgs(this) }
+
                 val subCommandLine = CommandLine(interactive)
                 println("You can run the same command again by using the following command line arguments:")
-                println("ccsh " + selectedParser + " " + collectedArgs.map { x -> '"' + x + '"' }.joinToString(" "))
+                println("ccsh " + selectedParser + " " + collectedArgs.joinToString(" ") { x -> '"' + x + '"' })
                 subCommandLine.execute(*collectedArgs.toTypedArray())
             } else {
                 printNotSupported(selectedParser)
@@ -80,7 +77,7 @@ class ParserService {
 
         private fun printNotSupported(parserName: String) {
             println(
-                "The interactive usage of $parserName is not supported yet.\n" + "Please specify the full command to run the parser."
+                "The interactive usage of $parserName is not supported yet.\nPlease specify the full command to run the parser."
             )
         }
     }